Make a more logical code structure
