+++ evio.go
@@ -9,6 +9,7 @@ import (
 	"net"
 	"os"
 	"strings"
+	"syscall"
 	"time"
 )
 
@@ -226,6 +227,47 @@ type listener struct {
 	fd      int
 	network string
 	addr    string
+}
+
+func (ln *listener) close() {
+	if ln.fd != 0 {
+		syscall.Close(ln.fd)
+	}
+	if ln.f != nil {
+		ln.f.Close()
+	}
+	if ln.ln != nil {
+		ln.ln.Close()
+	}
+	if ln.pconn != nil {
+		ln.pconn.Close()
+	}
+	if ln.network == "unix" {
+		os.RemoveAll(ln.addr)
+	}
+}
+
+// system takes the net listener and detaches it from it's parent
+// event loop, grabs the file descriptor, and makes it non-blocking.
+func (ln *listener) system() error {
+	var err error
+	switch netln := ln.ln.(type) {
+	case nil:
+		switch pconn := ln.pconn.(type) {
+		case *net.UDPConn:
+			ln.f, err = pconn.File()
+		}
+	case *net.TCPListener:
+		ln.f, err = netln.File()
+	case *net.UnixListener:
+		ln.f, err = netln.File()
+	}
+	if err != nil {
+		ln.close()
+		return err
+	}
+	ln.fd = int(ln.f.Fd())
+	return syscall.SetNonblock(ln.fd, true)
 }
 
 type addrOpts struct {
