+++ evio_unix.go
@@ -9,7 +9,6 @@ package evio
 import (
 	"io"
 	"net"
-	"os"
 	"runtime"
 	"sync"
 	"sync/atomic"
@@ -17,7 +16,7 @@ import (
 	"time"
 
 	reuseport "github.com/kavu/go_reuseport"
-	"github.com/tidwall/evio/internal"
+	"github.com/panjf2000/evio/internal"
 )
 
 type conn struct {
@@ -95,8 +94,8 @@ func serve(events Events, listeners []*listener) error {
 	s.cond = sync.NewCond(&sync.Mutex{})
 	s.balance = events.LoadBalance
 	s.tch = make(chan time.Duration)
-	s.fdlns = make(map[int]*listener, len(listeners))
 
+	s.fdlns = make(map[int]*listener, len(listeners))
 	for _, ln := range listeners {
 		s.fdlns[ln.fd] = ln
 	}
@@ -478,47 +477,6 @@ func (c *detachedConn) Write(p []byte) (n int, err error) {
 		p = p[nn:]
 	}
 	return n, nil
-}
-
-func (ln *listener) close() {
-	if ln.fd != 0 {
-		syscall.Close(ln.fd)
-	}
-	if ln.f != nil {
-		ln.f.Close()
-	}
-	if ln.ln != nil {
-		ln.ln.Close()
-	}
-	if ln.pconn != nil {
-		ln.pconn.Close()
-	}
-	if ln.network == "unix" {
-		os.RemoveAll(ln.addr)
-	}
-}
-
-// system takes the net listener and detaches it from it's parent
-// event loop, grabs the file descriptor, and makes it non-blocking.
-func (ln *listener) system() error {
-	var err error
-	switch netln := ln.ln.(type) {
-	case nil:
-		switch pconn := ln.pconn.(type) {
-		case *net.UDPConn:
-			ln.f, err = pconn.File()
-		}
-	case *net.TCPListener:
-		ln.f, err = netln.File()
-	case *net.UnixListener:
-		ln.f, err = netln.File()
-	}
-	if err != nil {
-		ln.close()
-		return err
-	}
-	ln.fd = int(ln.f.Fd())
-	return syscall.SetNonblock(ln.fd, true)
 }
 
 func reuseportListenPacket(proto, addr string) (l net.PacketConn, err error) {
